@@ -52,7 +52,11 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
 
     // Attempt repository path contraction (if we are in a git repository)
     // Otherwise use the logical path, automatically contracting
-    let repo = context.get_repo().ok();
+    let repo = if config.truncate_to_repo || config.repo_root_style.is_some() {
+        context.get_repo().ok()
+    } else {
+        None
+    };
     let dir_string = if config.truncate_to_repo {
         repo.and_then(|r| r.workdir.as_ref())
             .filter(|&root| root != &home_dir)
@@ -97,7 +101,7 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
             String::from(config.truncation_symbol)
         }
     } else {
-        String::from("")
+        String::new()
     };
 
     let path_vec = match &repo.and_then(|r| r.workdir.as_ref()) {
@@ -114,10 +118,10 @@ pub fn module<'a>(context: &'a Context) -> Option<Module<'a>> {
                 let before = before_root_dir(&dir_string, &contracted_path);
                 [prefix + before.as_str(), root.to_string(), after_repo_root]
             } else {
-                ["".to_string(), "".to_string(), prefix + dir_string.as_str()]
+                [String::new(), String::new(), prefix + dir_string.as_str()]
             }
         }
-        _ => ["".to_string(), "".to_string(), prefix + dir_string.as_str()],
+        _ => [String::new(), String::new(), prefix + dir_string.as_str()],
     };
 
     let path_vec = if config.use_os_path_sep {
@@ -179,7 +183,7 @@ fn remove_extended_path_prefix(path: String) -> String {
     }
     // Trim any Windows extended-path prefix from the display path
     if let Some(unc) = try_trim_prefix(&path, r"\\?\UNC\") {
-        return format!(r"\\{}", unc);
+        return format!(r"\\{unc}");
     }
     if let Some(p) = try_trim_prefix(&path, r"\\?\") {
         return p.to_string();
@@ -324,7 +328,7 @@ fn to_fish_style(pwd_dir_length: usize, dir_string: String, truncated_dir_string
         .map(|word| -> String {
             let chars = UnicodeSegmentation::graphemes(word, true).collect::<Vec<&str>>();
             match word {
-                "" => "".to_string(),
+                "" => String::new(),
                 _ if chars.len() <= pwd_dir_length => word.to_string(),
                 _ if word.starts_with('.') => chars[..=pwd_dir_length].join(""),
                 _ => chars[..pwd_dir_length].join(""),
@@ -497,7 +501,7 @@ mod tests {
 
     fn init_repo(path: &Path) -> io::Result<()> {
         create_command("git")?
-            .args(&["init"])
+            .args(["init"])
             .current_dir(path)
             .output()
             .map(|_| ())
@@ -685,7 +689,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("/foo/bar/{}/path", strange_sub)))
+                .paint(convert_path_sep(&format!("/foo/bar/{strange_sub}/path")))
         ));
 
         assert_eq!(expected, actual);
@@ -702,7 +706,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("~/{}/starship", name)))
+                .paint(convert_path_sep(&format!("~/{name}/starship")))
         ));
 
         assert_eq!(expected, actual);
@@ -720,7 +724,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/engine/schematics", name)))
+                .paint(convert_path_sep(&format!("{name}/engine/schematics")))
         ));
 
         assert_eq!(expected, actual);
@@ -784,7 +788,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/thrusters/rocket", name)))
+                .paint(convert_path_sep(&format!("{name}/thrusters/rocket")))
         ));
 
         assert_eq!(expected, actual);
@@ -860,7 +864,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("{}/rocket", name)))
+                .paint(convert_path_sep(&format!("{name}/rocket")))
         ));
 
         assert_eq!(expected, actual);
@@ -1113,7 +1117,7 @@ mod tests {
         let src_dir = repo_dir.join("src");
         let symlink_dir = tmp_dir.path().join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1139,7 +1143,7 @@ mod tests {
         let src_dir = repo_dir.join("src/meters/fuel-gauge");
         let symlink_dir = tmp_dir.path().join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1168,7 +1172,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1203,7 +1207,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1240,7 +1244,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1285,7 +1289,7 @@ mod tests {
             .join("above-repo")
             .join("rocket-controls-symlink");
         let symlink_src_dir = symlink_dir.join("src/meters/fuel-gauge");
-        fs::create_dir_all(&src_dir)?;
+        fs::create_dir_all(src_dir)?;
         init_repo(&repo_dir).unwrap();
         symlink(&repo_dir, &symlink_dir)?;
 
@@ -1395,7 +1399,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("…/{}/a/subpath", name)))
+                .paint(convert_path_sep(&format!("…/{name}/a/subpath")))
         ));
         assert_eq!(expected, actual);
         tmp_dir.close()
@@ -1420,7 +1424,7 @@ mod tests {
             "{} ",
             Color::Cyan
                 .bold()
-                .paint(convert_path_sep(&format!("~/{}/a/subpath", name)))
+                .paint(convert_path_sep(&format!("~/{name}/a/subpath")))
         ));
         assert_eq!(expected, actual);
         tmp_dir.close()
@@ -1790,7 +1794,7 @@ mod tests {
             .collect();
         let expected = Some(format!(
             "{} ",
-            Color::Cyan.bold().paint(format!("~/{}/starship", name))
+            Color::Cyan.bold().paint(format!("~/{name}/starship"))
         ));
 
         assert_eq!(expected, actual);
