@@ -194,7 +194,7 @@ mod test {
             .collect();
         let expected = Some(format!(
             "with {} ",
-            style().paint(format!("■ {}", TEST_VAR_VALUE))
+            style().paint(format!("■ {TEST_VAR_VALUE}"))
         ));
 
         assert_eq!(expected, actual);
@@ -211,7 +211,7 @@ mod test {
             .collect();
         let expected = Some(format!(
             "with {} ",
-            style().paint(format!("_{}", TEST_VAR_VALUE))
+            style().paint(format!("_{TEST_VAR_VALUE}"))
         ));
 
         assert_eq!(expected, actual);
@@ -228,7 +228,7 @@ mod test {
             .collect();
         let expected = Some(format!(
             "with {} ",
-            style().paint(format!("{}_", TEST_VAR_VALUE))
+            style().paint(format!("{TEST_VAR_VALUE}_"))
         ));
 
         assert_eq!(expected, actual);
