@@ -269,17 +269,17 @@ mod tests {
         let repo_dir = tempfile::tempdir()?;
 
         create_command("git")?
-            .args(&["init"])
+            .args(["init"])
             .current_dir(&repo_dir)
             .output()?;
 
         create_command("git")?
-            .args(&["symbolic-ref", "HEAD", "refs/heads/main"])
+            .args(["symbolic-ref", "HEAD", "refs/heads/main"])
             .current_dir(&repo_dir)
             .output()?;
 
         let actual = ModuleRenderer::new("git_branch")
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .collect();
 
         let expected = Some(format!(
@@ -296,7 +296,7 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "-b", "test_branch"])
+            .args(["checkout", "-b", "test_branch"])
             .current_dir(repo_dir.path())
             .output()?;
 
@@ -305,7 +305,7 @@ mod tests {
                 [git_branch]
                     only_attached = true
             })
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .collect();
 
         let expected = Some(format!(
@@ -324,8 +324,8 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "@~1"])
-            .current_dir(&repo_dir.path())
+            .args(["checkout", "@~1"])
+            .current_dir(repo_dir.path())
             .output()?;
 
         let actual = ModuleRenderer::new("git_branch")
@@ -333,7 +333,7 @@ mod tests {
                 [git_branch]
                     only_attached = true
             })
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .collect();
 
         let expected = None;
@@ -347,17 +347,17 @@ mod tests {
         let repo_dir = tempfile::tempdir()?;
 
         create_command("git")?
-            .args(&["init", "--bare"])
+            .args(["init", "--bare"])
             .current_dir(&repo_dir)
             .output()?;
 
         create_command("git")?
-            .args(&["symbolic-ref", "HEAD", "refs/heads/main"])
+            .args(["symbolic-ref", "HEAD", "refs/heads/main"])
             .current_dir(&repo_dir)
             .output()?;
 
         let actual = ModuleRenderer::new("git_branch")
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .collect();
 
         let expected = Some(format!(
@@ -374,7 +374,7 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "-b", "test_branch"])
+            .args(["checkout", "-b", "test_branch"])
             .current_dir(repo_dir.path())
             .output()?;
 
@@ -383,7 +383,7 @@ mod tests {
                 [git_branch]
                     ignore_branches = ["dummy", "test_branch"]
             })
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .collect();
 
         let expected = None;
@@ -398,23 +398,23 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "-b", "test_branch"])
+            .args(["checkout", "-b", "test_branch"])
             .current_dir(repo_dir.path())
             .output()?;
 
         create_command("git")?
-            .args(&["remote", "add", "--fetch", "remote_repo"])
+            .args(["remote", "add", "--fetch", "remote_repo"])
             .arg(remote_dir.path())
             .current_dir(repo_dir.path())
             .output()?;
 
         create_command("git")?
-            .args(&["branch", "--set-upstream-to", "remote_repo/master"])
+            .args(["branch", "--set-upstream-to", "remote_repo/master"])
             .current_dir(repo_dir.path())
             .output()?;
 
         let actual = ModuleRenderer::new("git_branch")
-            .path(&repo_dir.path())
+            .path(repo_dir.path())
             .config(toml::toml! {
                 [git_branch]
                 format = "$branch(:$remote_name/$remote_branch)"
@@ -482,7 +482,7 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "-b", branch_name])
+            .args(["checkout", "-b", branch_name])
             .current_dir(repo_dir.path())
             .output()?;
 
@@ -505,7 +505,7 @@ mod tests {
             "on {} ",
             Color::Purple
                 .bold()
-                .paint(format!("\u{e0a0} {}{}", expected_name, truncation_symbol)),
+                .paint(format!("\u{e0a0} {expected_name}{truncation_symbol}")),
         ));
 
         assert_eq!(expected, actual);
@@ -521,7 +521,7 @@ mod tests {
         let repo_dir = fixture_repo(FixtureProvider::Git)?;
 
         create_command("git")?
-            .args(&["checkout", "-b", branch_name])
+            .args(["checkout", "-b", branch_name])
             .current_dir(repo_dir.path())
             .output()?;
 
