@@ -270,6 +270,17 @@ active boot switches: -d:release\n",
             stdout: String::from("4.10.0\n"),
             stderr: String::default(),
         }),
+        "opa version" => Some(CommandOutput {
+            stdout: String::from("Version: 0.44.0
+Build Commit: e8d488f
+Build Timestamp: 2022-09-07T23:50:25Z
+Build Hostname: 119428673f4c
+Go Version: go1.19.1
+Platform: linux/amd64
+WebAssembly: unavailable
+"),
+            stderr: String::default(),
+        }),
         "opam switch show --safe" => Some(CommandOutput {
             stdout: String::from("default\n"),
             stderr: String::default(),
@@ -422,17 +433,17 @@ pub fn wrap_seq_for_shell(
             if x == escape_begin && !escaped {
                 escaped = true;
                 match shell {
-                    Shell::Bash => format!("{}{}", BASH_BEG, escape_begin),
-                    Shell::Zsh => format!("{}{}", ZSH_BEG, escape_begin),
-                    Shell::Tcsh => format!("{}{}", TCSH_BEG, escape_begin),
+                    Shell::Bash => format!("{BASH_BEG}{escape_begin}"),
+                    Shell::Zsh => format!("{ZSH_BEG}{escape_begin}"),
+                    Shell::Tcsh => format!("{TCSH_BEG}{escape_begin}"),
                     _ => x.to_string(),
                 }
             } else if x == escape_end && escaped {
                 escaped = false;
                 match shell {
-                    Shell::Bash => format!("{}{}", escape_end, BASH_END),
-                    Shell::Zsh => format!("{}{}", escape_end, ZSH_END),
-                    Shell::Tcsh => format!("{}{}", escape_end, TCSH_END),
+                    Shell::Bash => format!("{escape_end}{BASH_END}"),
+                    Shell::Zsh => format!("{escape_end}{ZSH_END}"),
+                    Shell::Tcsh => format!("{escape_end}{TCSH_END}"),
                     _ => x.to_string(),
                 }
             } else {
@@ -548,7 +559,7 @@ pub fn render_time(raw_millis: u128, show_millis: bool) -> String {
 fn render_time_component((component, suffix): (&u128, &&str)) -> String {
     match component {
         0 => String::new(),
-        n => format!("{}{}", n, suffix),
+        n => format!("{n}{suffix}"),
     }
 }
 
@@ -739,7 +750,7 @@ mod tests {
         };
         assert_eq!(get_command_string_output(case1), "stdout");
         let case2 = CommandOutput {
-            stdout: String::from(""),
+            stdout: String::new(),
             stderr: String::from("stderr"),
         };
         assert_eq!(get_command_string_output(case2), "stderr");
