@@ -23,7 +23,7 @@ use std::collections::HashMap;
 /// When using [awsu](https://github.com/kreuzwerker/awsu) the profile
 /// is read from the `AWSU_PROFILE` env var.
 ///
-/// When using [AWSume](https://awsu.me) the profile
+/// When using [`AWSume`](https://awsu.me) the profile
 /// is read from the `AWSUME_PROFILE` env var and the credentials expiration
 /// date is read from the `AWSUME_EXPIRATION` env var.
 pub struct AwsConfig<'a> {
