@@ -103,7 +103,7 @@ mod tests {
                 .sync_all()
                 .unwrap();
             let actual = ModuleRenderer::new("buf").path(dir.path()).collect();
-            let expected = Some(format!("with {}", Color::Blue.bold().paint(" v1.0.0")));
+            let expected = Some(format!("with {}", Color::Blue.bold().paint("🦬 v1.0.0 ")));
             assert_eq!(expected, actual);
             dir.close().unwrap();
         }
