@@ -145,7 +145,7 @@ mod tests {
     fn folder_with_godeps() -> io::Result<()> {
         let dir = tempfile::tempdir()?;
         let godeps = dir.path().join("Godeps");
-        fs::create_dir_all(&godeps)?;
+        fs::create_dir_all(godeps)?;
 
         let actual = ModuleRenderer::new("golang").path(dir.path()).collect();
 
