@@ -118,7 +118,7 @@ mod tests {
     fn folder_with_lua_folder() -> io::Result<()> {
         let dir = tempfile::tempdir()?;
         let lua_dir = dir.path().join("lua");
-        fs::create_dir_all(&lua_dir)?;
+        fs::create_dir_all(lua_dir)?;
 
         let actual = ModuleRenderer::new("lua").path(dir.path()).collect();
         let expected = Some(format!("via {}", Color::Blue.bold().paint("🌙 v5.4.0 ")));
