@@ -30,6 +30,7 @@ mod git_metrics;
 mod git_state;
 mod git_status;
 mod golang;
+mod guix_shell;
 mod haskell;
 mod helm;
 mod hg_branch;
@@ -43,11 +44,14 @@ mod line_break;
 mod localip;
 mod lua;
 mod memory_usage;
+mod meson;
 mod nim;
 mod nix_shell;
 mod nodejs;
 mod ocaml;
+mod opa;
 mod openstack;
+mod os;
 mod package;
 mod perl;
 mod php;
@@ -124,6 +128,7 @@ pub fn handle<'a>(module: &str, context: &'a Context) -> Option<Module<'a>> {
             "git_state" => git_state::module(context),
             "git_status" => git_status::module(context),
             "golang" => golang::module(context),
+            "guix_shell" => guix_shell::module(context),
             "haskell" => haskell::module(context),
             "helm" => helm::module(context),
             "hg_branch" => hg_branch::module(context),
@@ -137,11 +142,14 @@ pub fn handle<'a>(module: &str, context: &'a Context) -> Option<Module<'a>> {
             "localip" => localip::module(context),
             "lua" => lua::module(context),
             "memory_usage" => memory_usage::module(context),
+            "meson" => meson::module(context),
             "nim" => nim::module(context),
             "nix_shell" => nix_shell::module(context),
             "nodejs" => nodejs::module(context),
             "ocaml" => ocaml::module(context),
+            "opa" => opa::module(context),
             "openstack" => openstack::module(context),
+            "os" => os::module(context),
             "package" => package::module(context),
             "perl" => perl::module(context),
             "php" => php::module(context),
@@ -176,7 +184,7 @@ pub fn handle<'a>(module: &str, context: &'a Context) -> Option<Module<'a>> {
                 custom::module(custom.strip_prefix("custom.").unwrap(), context)
             }
             _ => {
-                eprintln!("Error: Unknown module {}. Use starship module --list to list out all supported modules.", module);
+                eprintln!("Error: Unknown module {module}. Use starship module --list to list out all supported modules.");
                 None
             }
         }
@@ -229,6 +237,7 @@ pub fn description(module: &str) -> &'static str {
         "git_state" => "The current git operation, and it's progress",
         "git_status" => "Symbol representing the state of the repo",
         "golang" => "The currently installed version of Golang",
+        "guix_shell" => "The guix-shell environment",
         "haskell" => "The selected version of the Haskell toolchain",
         "helm" => "The currently installed version of Helm",
         "hg_branch" => "The active branch of the repo in your current directory",
@@ -242,11 +251,16 @@ pub fn description(module: &str) -> &'static str {
         "localip" => "The currently assigned ipv4 address",
         "lua" => "The currently installed version of Lua",
         "memory_usage" => "Current system memory and swap usage",
+        "meson" => {
+            "The current Meson environment, if $MESON_DEVENV and $MESON_PROJECT_NAME are set"
+        }
         "nim" => "The currently installed version of Nim",
         "nix_shell" => "The nix-shell environment",
         "nodejs" => "The currently installed version of NodeJS",
         "ocaml" => "The currently installed version of OCaml",
+        "opa" => "The currently installed version of Open Platform Agent",
         "openstack" => "The current OpenStack cloud and project",
+        "os" => "The current operating system",
         "package" => "The package version of the current directory's project",
         "perl" => "The currently installed version of Perl",
         "php" => "The currently installed version of PHP",
@@ -285,7 +299,7 @@ mod test {
     #[test]
     fn all_modules_have_description() {
         for module in ALL_MODULES {
-            println!("Checking if {:?} has a description", module);
+            println!("Checking if {module:?} has a description");
             assert_ne!(description(module), "<no description>");
         }
     }
