@@ -123,7 +123,7 @@ fn find_package_file(path: &Path) -> Option<PathBuf> {
 /// Pulumi has no CLI option that is fast enough to get this for us, but finding
 /// the location is simple. We get it ourselves.
 fn stack_name(project_file: &Path, context: &Context) -> Option<String> {
-    let mut file = File::open(&project_file).ok()?;
+    let mut file = File::open(project_file).ok()?;
 
     let mut contents = String::new();
     file.read_to_string(&mut contents).ok()?;
@@ -169,7 +169,7 @@ fn get_pulumi_workspace(context: &Context, name: &str, project_file: &Path) -> O
         hasher.update(project_file.to_str()?.as_bytes());
         crate::utils::encode_to_hex(&hasher.finalize())
     };
-    let unique_file_name = format!("{}-{}-workspace.json", name, project_file);
+    let unique_file_name = format!("{name}-{project_file}-workspace.json");
     let mut path = pulumi_home_dir(context)?;
     path.push("workspaces");
     path.push(unique_file_name);
@@ -287,7 +287,7 @@ mod tests {
         let workspace_path = root.join(".pulumi").join("workspaces");
         std::fs::create_dir_all(&workspace_path)?;
         let workspace_path = &workspace_path.join("starship-test-workspace.json");
-        let mut workspace = File::create(&workspace_path)?;
+        let mut workspace = File::create(workspace_path)?;
         serde_json::to_writer_pretty(
             &mut workspace,
             &serde_json::json!(
@@ -301,7 +301,7 @@ mod tests {
         let credential_path = root.join(".pulumi");
         std::fs::create_dir_all(&credential_path)?;
         let credential_path = &credential_path.join("credentials.json");
-        let mut credential = File::create(&credential_path)?;
+        let mut credential = File::create(credential_path)?;
         serde_json::to_writer_pretty(
             &mut credential,
             &serde_json::json!(
@@ -354,7 +354,7 @@ mod tests {
         let workspace_path = root.join(".pulumi").join("workspaces");
         std::fs::create_dir_all(&workspace_path)?;
         let workspace_path = &workspace_path.join("starship-test-workspace.json");
-        let mut workspace = File::create(&workspace_path)?;
+        let mut workspace = File::create(workspace_path)?;
         serde_json::to_writer_pretty(
             &mut workspace,
             &serde_json::json!(
@@ -368,7 +368,7 @@ mod tests {
         let credential_path = root.join(".pulumi");
         std::fs::create_dir_all(&credential_path)?;
         let credential_path = &credential_path.join("starship-test-credential.json");
-        let mut credential = File::create(&credential_path)?;
+        let mut credential = File::create(credential_path)?;
         serde_json::to_writer_pretty(
             &mut credential,
             &serde_json::json!(
